Fixed unit test bugs
Was allowing asset IDs my tests were set up to disallow.

Made the test asset IDs hardcoded.  This allows for easier "check the filesystem" type debugging.

Fixed the cache path generation to match Chattel.  Yep, it's fragile code.

Foudn that request.RawBytes is actually the bytes of the content, exactly what I needed, not the bytes of the whole request with headers like I thought it was.  Trying to decode the string data was leading to big problems with mismatches on certain characters.

Fixed the TGA test to actually request a TGA asset...